requirements.txt
Flask==3.1.0
mysql_connector_repackaged==0.3.1
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.36
